djangolab/learning/requirements.txt
Django==3.2
django-rest-framework==0.1.0
joblib==1.0.1
scikit-learn==0.24.1